Moved classes to libgpml, added Resources.java back
Moved classes to libgpml --  in Pathway.java and PathwayElement.java there are specific links to PreferenceManager, BatikImageExporter, and VState. These are commented out because otherwise we cannot separate Model from View. Will handle these later.

Resources.java is added back to Utils and put back into pathvisio.core because the resources folder is in pathvisio.core and not in pathvisio.libgpml. The merging of resources and util may not be practical at this point.