Temporary fix Line issue for fireObjectModifiedEvent() PathwayObject/Listener Issue
In the past GenericPoint/LinePoint were not themselves PathwayElements and therefore when  fireObjectModifiedEvent is called, there is no confusion as to whether LinePoint or the parent LineElement is modified.

To bypass this issue: when calling fireObjectModifiedEvent  from a LinePoint method, we must make sure to call like this "LineElement.this.fireObjectModifiedEvent(				PathwayObjectEvent.createSinglePropertyEvent(LineElement.this, StaticProperty....));"

Should pay close attention to how things have changed. As It is no longer PathwayElementEvent but now PathwayObject Event.